requirements.txt
# Optional dependency for FAISS-based negative mining
faiss-cpu>=1.7.4

